+++ src/jlmkr/actions/edit.py
@@ -7,8 +7,8 @@ import subprocess
 
 from utils.console import eprint
 from utils.editor import get_text_editor
-from utils.jail_dataset import check_jail_name_valid, check_jail_name_available
-from utils.jail_dataset import get_jail_config_path, jail_is_running
+from utils.dataset import check_jail_name_valid, check_jail_name_available
+from utils.dataset import get_jail_config_path, jail_is_running
 
 
 def edit_jail(jail_name):
